+++ CHANGELOG.md
@@ -15,6 +15,7 @@
     * Fixed issue with duplicate configuration settings in Flex Directory
     * Fixed fatal error if there are numeric folders in `Flex Pages`
     * Fixed error on missing `Flex` templates in if `Flex Objects` plugin isn't installed
+    * Fixed `PageTranslateTrait::getAllLanguages()` and `getAllLanguages()` to include default language
     * Grav 1.7: Fixed PHP 7.1 compatibility issues
     * Selfupgrade CLI: Fixed broken selfupgrade assets reference [#2681](https://github.com/getgrav/grav/issues/2681)